BUG/MINOR: otel: cast the name byte to uint8_t for the ctype check
The variable-name check passed each name byte, a plain (signed) char, to
isdigit() and isalnum().  A byte with the high bit set promotes to a
negative int, which is undefined for the ctype functions: the argument
must be representable as an unsigned char or equal EOF.  Each byte is now
cast to uint8_t first, matching the uint8_t spelling already used in the
name normaliser and at HAProxy's own ctype call sites.

+++ src/parser.c
@@ -154,8 +154,8 @@ static const char *flt_otel_parse_invalid_char(const char *name, int type)
 		 * Allowed characters are letters, numbers and '_', the first
 		 * character in the string must not be a number.
 		 */
-		if (!isdigit(*retptr))
-			for (++retptr; (*retptr == '_') || isalnum(*retptr); retptr++);
+		if (!isdigit((uint8_t)*retptr))
+			for (++retptr; (*retptr == '_') || isalnum((uint8_t)*retptr); retptr++);
 
 		if (*retptr == '\0')
 			retptr = NULL;
